sneaking
I struggled with this for a couple hours, then after going back to the original, changing from CanSee to IsSneaking because I thought stating with a false value might initialize better. None of that made any difference until I added a Wait task to the far right. This seems to give the tree a chance to think and reset the value for IsSneaking. Anyway it seems to work now. I also took out the issneaking  decorator on ranged attacks just to simplify debugging. We can add that back later.